Move views to port 80
